requirements.txt
aiogram==3.0.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
python-dotenv==1.0.0
stripe==8.0.0
paypalrestsdk==1.13.1
googletrans==4.0.0-rc1
openai==1.0.0
apscheduler==3.10.0
aiohttp==3.9.0
pandas==2.1.0
Pillow==10.0.0
matplotlib==3.7.0
scikit-learn==1.3.0
numpy==1.26.0